requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#
altair==5.3.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   groq
    #   httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via newspaper4k
blinker==1.8.1
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.2.2
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via curl-cffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   nltk
    #   streamlit
    #   typer
curl-cffi==0.6.3
    # via duckduckgo-search
distro==1.9.0
    # via groq
duckduckgo-search==5.3.0
exceptiongroup==1.2.1
    # via anyio
feedparser==6.0.11
    # via newspaper4k
filelock==3.14.0
    # via tldextract
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   phidata
    #   streamlit
groq==0.5.0
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   groq
    #   phidata
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   tldextract
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
joblib==1.4.0
    # via nltk
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
lxml==5.2.1
    # via
    #   lxml-html-clean
    #   newspaper4k
lxml-html-clean==0.1.1

markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.6.0

newspaper4k==0.9.3.1

nltk==3.8.1
    # via newspaper4k
numpy==1.26.4
    # via
    #   altair
    #   newspaper4k
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
orjson==3.10.1
    # via duckduckgo-search
packaging==24.0
    # via
    #   altair
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   newspaper4k
    #   streamlit
phidata==2.3.89

pillow==10.3.0
    # via
    #   newspaper4k
    #   streamlit
protobuf==4.25.3
    # via streamlit
pyarrow==16.0.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   groq
    #   phidata
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via phidata
pydeck==0.9.0
    # via streamlit
pygments==2.17.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   newspaper4k
    #   pandas
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   newspaper4k
    #   phidata
referencing==0.35.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.4.28
    # via nltk
requests==2.31.0
    # via
    #   newspaper4k
    #   requests-file
    #   streamlit
    #   tldextract
requests-file==2.0.0
    # via tldextract
rich==13.7.1
    # via
    #   phidata
    #   streamlit
    #   typer
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
streamlit==1.33.0

tenacity==8.2.3
    # via streamlit
tldextract==5.1.2
    # via newspaper4k
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4
    # via streamlit
tqdm==4.66.2
    # via nltk
typer==0.12.3
    # via phidata
typing-extensions==4.11.0
    # via
    #   altair
    #   anyio
    #   groq
    #   phidata
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests